hook: save price to level db instead of price cache
